move pando Document react components
