+++ libs/pando/ui-sheet/src/index.ts
@@ -1,5 +1,3 @@
-export * from './conditional'
-export * from './document'
-export * from './field'
-export * from './header'
-export * from './sheet'
+export * from './components'
+export * from './context'
+export * from './types'
